+++ zlt-uaa/src/main/java/com/central/oauth/openid/OpenIdAuthenticationProvider.java
@@ -1,9 +1,9 @@
 package com.central.oauth.openid;
 
+import com.central.oauth2.common.token.OpenIdAuthenticationToken;
 import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.authentication.InternalAuthenticationServiceException;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.social.security.SocialUserDetailsService;
